move: 레스토랑 domain 객체 이름 변경
