+++ src/domain/RestaurantCollector.ts
