+++ src/index.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 
 import App from './App';
-import RestaurantManager from './domain/RestaurantManager';
+import RestaurantManager from './domain/RestaurantCollector';
 import { GlobalStyle } from './globalStyle';
 
 RestaurantManager.initMockData();
